fix(EditableTable): cell icon/placeholder alignment and disabled-cell icons
- Lift the input's leading icon by 3px and the placeholder by 1px inside
  OneEditableTable cells so both sit on the cell's vertical center.
- Read-only select cells reuse F0Select's Arrow (was a smaller F0Icon), so the
  dropdown chevron matches the editable select cell instead of looking scaled
  down.
- Disabled cells now show the same field affordances as other cells (e.g. the
  url/email leading icon), just muted — via a shared ReadOnlyCellContent used by
  both the display-only and disabled cells.
