+++ packages/react/src/patterns/OneDataCollection/visualizations/collection/EditableTable/components/cells/status/DisabledCell.tsx
@@ -1,18 +1,14 @@
 import { RecordType } from "@/hooks/datasource/types/records.typings"
-import { useI18n } from "@/lib/providers/i18n/i18n-provider"
-import { cn } from "@/lib/utils"
-import { renderProperty } from "@/patterns/OneDataCollection/property-render"
 
 import { EditableCellProps } from ".."
 import { BaseCell } from "../BaseCell"
+import { ReadOnlyCellContent } from "../ReadOnlyCellContent"
 
 export function DisabledCell<R extends RecordType>({
   editableColumn,
   item,
   hint,
 }: EditableCellProps<R>) {
-  const i18n = useI18n()
-
   return (
     <BaseCell
       borderOnHover={false}
@@ -20,19 +16,14 @@ export function DisabledCell<R extends RecordType>({
       hintPosition="right"
       cursor="not-allowed"
     >
-      <div
-        className={cn(
-          editableColumn.align === "right" ? "justify-end" : "",
-          // px only (no vertical padding) so the cell matches the 48px height
-          // of editable/display-only cells and rows stay vertically aligned.
-          "flex px-3 min-h-12 items-center border-0 h-full",
-          "bg-f1-background-disabled h-full",
-          "w-full",
-          "[&_*]:text-f1-foreground-secondary"
-        )}
-      >
-        {renderProperty(item, editableColumn, "editableTable", i18n)}
-      </div>
+      <ReadOnlyCellContent
+        editableColumn={editableColumn}
+        item={item}
+        iconColor="secondary"
+        // px only (no vertical padding) so the disabled cell matches the 48px
+        // height of editable/display-only cells and rows stay aligned.
+        className="min-h-12 bg-f1-background-disabled [&_*]:text-f1-foreground-secondary"
+      />
     </BaseCell>
   )
 }
